+++ packages/patterns-react/src/patterns/SimpleLongForm/SimpleLongForm.js
@@ -11,7 +11,7 @@ import classNames from 'classnames';
 import { settings as ddsSettings } from '@carbon/ibmdotcom-utilities';
 import { settings } from 'carbon-components';
 import { featureFlag } from '@carbon/ibmdotcom-utilities';
-import { SIMPLELONGFORM } from '../../internal/FeatureFlags';
+import { DDS_SIMPLELONGFORM } from '../../internal/FeatureFlags';
 import { ArrowRight20 } from '@carbon/icons-react';
 import { CardLink } from '@carbon/ibmdotcom-react';
 import { LinkWithIcon } from '@carbon/ibmdotcom-react';
@@ -23,28 +23,33 @@ const { prefix } = settings;
  * Simple long form pattern
  *
  * @param {object} props props object
- * @param {string} props.title simple long form title
+ * @param {boolean} props.border includes border or not ( true | false )
  * @param {string} props.copy simple long form  short copy to support the title
  * @param {string} props.linkType link type ( simple | jump | card )
  * @param {object} props.link link object which includes url, link text and target properties.
+ * @param {string} props.theme simple long form theme (g100 | white/default)
+ * @param {string} props.title simple long form title
  * @returns {*} Simple long form pattern
  */
-const SimpleLongForm = ({ title, copy, linkType, border, link }) =>
+const SimpleLongForm = ({ border, copy, linkType, link, theme, title }) =>
   featureFlag(
-    SIMPLELONGFORM,
+    DDS_SIMPLELONGFORM,
     <section
       data-autoid={`${stablePrefix}--simplelongform`}
       className={classNames(
         `${prefix}--simplelongform`,
         setLinkType(linkType),
-        setBorder(border)
+        setBorder(border),
+        setTheme(theme)
       )}>
       <div className={`${prefix}--simplelongform__container`}>
         <div className={`${prefix}--simplelongform__row`}>
           <div className={`${prefix}--simplelongform__col`}>
             <h3 className={`${prefix}--simplelongform__title`}>{title}</h3>
             <div className={`${prefix}--simplelongform__content`}>{copy}</div>
-            <div className={`${prefix}--simplelongform__link`}>
+            <div
+              data-autoid={`${stablePrefix}--simplelongform__link`}
+              className={`${prefix}--simplelongform__link`}>
               {renderLink(linkType, link)}
             </div>
           </div>
@@ -102,6 +107,16 @@ const setLinkType = type => {
 };
 
 /**
+ * sets the class name based on theme type
+ *
+ * @param {string} theme theme type ( g100 | white/default )
+ * @returns {string} theme css class names
+ */
+const setTheme = theme => {
+  return theme && `${prefix}--simplelongform--${theme}`;
+};
+
+/**
  * sets the class name based on border type
  *
  * @param {boolean} border includes border or not ( true | false )
@@ -114,15 +129,16 @@ const setBorder = border => {
 };
 
 SimpleLongForm.propTypes = {
-  title: PropTypes.string.isRequired,
-  copy: PropTypes.string,
-  linkType: PropTypes.string,
   border: PropTypes.bool,
+  copy: PropTypes.string,
   link: PropTypes.shape({
     href: PropTypes.string,
     text: PropTypes.string,
     target: PropTypes.string,
   }),
+  linkType: PropTypes.string,
+  theme: PropTypes.string,
+  title: PropTypes.string.isRequired,
 };
 
 export default SimpleLongForm;