+++ packages/react/src/components/ButtonGroup/__stories__/ButtonGroup.stories.js
@@ -2,13 +2,14 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text, select } from '@storybook/addon-knobs';
 import { ArrowRight20, ArrowDown20, Pdf20 } from '@carbon/icons-react';
+import '@carbon/grid/scss/grid.scss';
 import '../../../../../styles/scss/components/buttongroup/_buttongroup.scss';
 import readme from '../README.md';
-import { BUTTON_GROUP } from '../../../internal/FeatureFlags';
+import { DDS_BUTTON_GROUP } from '../../../internal/FeatureFlags';
 
 import ButtonGroup from '../ButtonGroup';
 
-if (BUTTON_GROUP) {
+if (DDS_BUTTON_GROUP) {
   storiesOf('ButtonGroup', module)
     .addDecorator(withKnobs)
     .addParameters({
@@ -54,10 +55,24 @@ if (BUTTON_GROUP) {
 
       return (
         <div
+          className="bx-grid"
           style={{
             padding: 2 + `rem`,
             backgroundColor: 'black',
           }}>
+          <div style={{ color: 'white' }}>
+            This button group is wrapped within the grid to let the buttons
+            shrink when the text gets smaller
+          </div>
+          <div className="bx--row">
+            <div className="bx--col-lg-16 bx--col-md-6 bx--col-sm-16">
+              <ButtonGroup buttons={buttons} />
+            </div>
+          </div>
+          <div style={{ color: 'white', paddingTop: '20px' }}>
+            This button group is not using the grid, so the buttons won't shrink
+            according to the text size
+          </div>
           <ButtonGroup buttons={buttons} />
         </div>
       );