+++ packages/react/src/components/CardLink/__stories__/Card.stories.js
@@ -4,11 +4,11 @@ import { withKnobs, select, text } from '@storybook/addon-knobs';
 import { CardLink } from '../';
 import { ArrowRight20 } from '@carbon/icons-react';
 import readme from '../README.md';
-import { CARD_LINK } from '../../../internal/FeatureFlags';
+import { DDS_CARD_LINK } from '../../../internal/FeatureFlags';
 
 import './index.scss';
 
-if (CARD_LINK) {
+if (DDS_CARD_LINK) {
   storiesOf('Card', module)
     .addDecorator(withKnobs)
     .addDecorator(Story => {