+++ packages/react/src/components/LinkWithIcon/LinkWithIcon.js
@@ -35,7 +35,7 @@ const LinkWithIcon = ({ children, href, ...props }) => {
  * @property propTypes
  * @description Defined property types for component
  *
- * @type {{children: object, href: string}}
+ * @type {{children: array, href: string}}
  */
 LinkWithIcon.propTypes = {
   children: PropTypes.array,
@@ -44,7 +44,7 @@ LinkWithIcon.propTypes = {
 
 /**
  * @property defaultProps
- * @type {{children: {}, href: string}}
+ * @type {{children: [], href: string}}
  */
 LinkWithIcon.defaultProps = {
   children: [],