+++ packages/react/src/components/Masthead/Masthead.js
@@ -10,7 +10,6 @@ import PropTypes from 'prop-types';
 import { settings as ddsSettings } from '@carbon/ibmdotcom-utilities';
 import { settings } from 'carbon-components';
 import root from 'window-or-global';
-import { globalInit } from '@carbon/ibmdotcom-utilities';
 import { User20, UserOnline20 } from '@carbon/icons-react';
 import { IbmLogo } from '../Icon';
 import {
@@ -20,14 +19,18 @@ import {
   HeaderGlobalBar,
   SkipToContent,
 } from 'carbon-components-react';
-import { ProfileAPI, TranslationAPI } from '@carbon/ibmdotcom-services';
+import {
+  globalInit,
+  ProfileAPI,
+  TranslationAPI,
+} from '@carbon/ibmdotcom-services';
 import MastheadL1 from './MastheadL1';
 import MastheadSearch from './MastheadSearch';
 import MastheadProfile from './MastheadProfile';
 import MastheadLeftNav from './MastheadLeftNav';
 import MastheadTopNav from './MastheadTopNav';
 import cx from 'classnames';
-import { MASTHEAD_L1 } from '../../internal/FeatureFlags';
+import { DDS_MASTHEAD_L1 } from '../../internal/FeatureFlags';
 
 const { stablePrefix } = ddsSettings;
 const { prefix } = settings;
@@ -35,8 +38,10 @@ const { prefix } = settings;
 /**
  * MastHead component
  *
- * @typedef {object} navigation Object containing navigation elements
- * @param {string} type Type of masthead
+ * @param {object} props React props object
+ * @param {object} props.navigation Object containing navigation elements
+ * @param {boolean} props.hasProfile Determines whether to render Profile component
+ * @param {boolean} props.hasSearch Determines whether to render Search Bar
  * @returns {*} Masthead component
  */
 const Masthead = ({ navigation, hasProfile, hasSearch, ...mastheadProps }) => {
@@ -169,6 +174,7 @@ const Masthead = ({ navigation, hasProfile, hasSearch, ...mastheadProps }) => {
                 {hasSearch && (
                   <MastheadSearch
                     searchOpenOnload={mastheadProps.searchOpenOnload}
+                    placeHolderText={mastheadProps.placeHolderText}
                   />
                 )}
               </div>
@@ -199,9 +205,14 @@ const Masthead = ({ navigation, hasProfile, hasSearch, ...mastheadProps }) => {
               )}
             </Header>
           </div>
-          {MASTHEAD_L1 && navigation && (
+          {DDS_MASTHEAD_L1 && navigation && (
             <div ref={mastheadL1Ref}>
-              <MastheadL1 />
+              <MastheadL1
+                isShort={isMastheadSticky}
+                title={mastheadProps.title}
+                eyebrowText={mastheadProps.eyebrowText}
+                eyebrowLink={mastheadProps.eyebrowLink}
+              />
             </div>
           )}
         </div>
@@ -214,13 +225,13 @@ const Masthead = ({ navigation, hasProfile, hasSearch, ...mastheadProps }) => {
  * @property propTypes
  * @description Defined property types for component
  *
- * @type {{mastheadProp: object, navigation: object, hasProfile: boolean, hasSearch: boolean}}
+ * @type {{mastheadProps: object, navigation: object, hasProfile: boolean, hasSearch: boolean}}
  */
 Masthead.propTypes = {
   navigation: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
   hasProfile: PropTypes.bool,
   hasSearch: PropTypes.bool,
-  mastheadProp: PropTypes.object,
+  mastheadProps: PropTypes.object,
 };
 
 /**