+++ packages/react/src/components/index.js
@@ -14,5 +14,7 @@ export * from './Masthead';
 export * from './LinkWithIcon';
 export * from './LocaleModal';
 export * from './CardLink';
+export * from './Layout';
 export * from './ExpressiveModal';
 export * from './carbon-components-react/UIShell';
+export * from './TableOfContents';