+++ packages/styles/.storybook/addons.js
@@ -1,0 +1,6 @@
+import '@storybook/addon-knobs/register';
+import '@storybook/addon-storysource/register';
+import '@storybook/addon-actions/register';
+import '@storybook/addon-links/register';
+import 'storybook-readme/register';
+// import '@storybook/addon-options/register';